IGNITE-25428 Move defragmentation logic to NodeFileTree
